requirements.txt
asgiref==3.9.1
Django==5.2.4
Django_GCC_PhoneNumbers==0.0.3
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
